requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Google Gemini
google-generativeai==0.5.4

# OCR
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0
opencv-python-headless==4.9.0.80

# Data validation
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
python-dateutil==2.9.0
regex==2024.4.28
loguru==0.7.2

# Redis (optional caching)
redis==5.0.4
hiredis==2.3.2

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
